requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
boto3==1.36.25
botocore==1.36.25
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
Django==5.1.6
django-cors-headers==4.7.0
django-extensions==3.2.3
django-fcm==0.1.1
django-filter==25.1
django-storages==1.14.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
fcm-django==2.2.1
firebase-admin==6.8.0
frozenlist==1.6.0
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
jmespath==1.0.1
msgpack==1.1.0
multidict==6.4.3
pillow==11.2.1
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyfcm==2.0.8
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.3
rsa==4.9.1
s3transfer==0.11.2
six==1.17.0
sqlparse==0.5.3
swapper==1.4.0
uritemplate==4.1.1
urllib3==2.3.0
yarl==1.20.0
